# lib/api-zod/src/generated/types/agingBucket.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AR CRM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AgingBucket {
  count: number;
  amount: number;
}