# lib/api-zod/src/generated/types/customerSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AR CRM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerSummary {
  customerId: string;
  customerName: string;
  outstanding: number;
  invoiceCount: number;
}