# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AR CRM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerSummary } from './customerSummary';

export interface DashboardSummary {
  totalOutstanding: number;
  totalInvoices: number;
  overdueCount: number;
  overdueAmount: number;
  currentCount: number;
  notDueCount: number;
  collectedThisMonth: number;
  dso: number;
  disputedCount: number;
  disputedAmount: number;
  ptpCount: number;
  ptpAmount: number;
  topCustomers: CustomerSummary[];
}