# lib/api-zod/src/generated/types/emailReplyInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AR CRM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmailReplyInputDirection } from './emailReplyInputDirection';

export interface EmailReplyInput {
  /** @minLength 1 */
  body: string;
  direction: EmailReplyInputDirection;
}