# lib/api-zod/src/generated/types/emailThread.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AR CRM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmailMessage } from './emailMessage';
import type { EmailThreadStatus } from './emailThreadStatus';

export interface EmailThread {
  id: number;
  subject: string;
  invoiceId: string;
  invoiceNumber: string;
  customerId: string;
  customerName: string;
  customerEmail?: string;
  analystId: number;
  analystName: string;
  status: EmailThreadStatus;
  createdAt: string;
  updatedAt: string;
  messages: EmailMessage[];
}