# lib/api-zod/src/generated/types/emailThreadInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AR CRM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmailThreadInput {
  invoiceId: string;
  analystId: number;
  to: string;
  body: string;
}